Extract DispositionService to centralize disposition/willingness logic
Move all disposition calculations (club willingness to sell, player
willingness to transfer/renew/sign) from TransferService, ContractService,
ScoutingService, and LoanService into a new DispositionService. This
eliminates duplicated mood indicator code and gives these core game
algorithms a single home. Original services retain thin delegation
wrappers to preserve existing call sites. No behavioral changes.
