+++ app/Http/Actions/NegotiatePreContract.php
@@ -7,6 +7,7 @@ use App\Models\GamePlayer;
 use App\Models\TransferOffer;
 use App\Modules\Notification\Services\NotificationService;
 use App\Modules\Transfer\Services\ContractService;
+use App\Modules\Transfer\Services\DispositionService;
 use App\Modules\Transfer\Services\ScoutingService;
 use App\Support\Money;
 use Illuminate\Http\JsonResponse;
@@ -21,6 +22,7 @@ class NegotiatePreContract
         private readonly ContractService $contractService,
         private readonly ScoutingService $scoutingService,
         private readonly NotificationService $notificationService,
+        private readonly DispositionService $dispositionService,
     ) {}
 
     public function __invoke(Request $request, string $gameId, string $playerId): JsonResponse
@@ -73,7 +75,7 @@ class NegotiatePreContract
             ->first();
 
         if ($existing) {
-            $mood = $this->getWillingnessMood($player, $game);
+            $mood = $this->dispositionService->willingnessMoodIndicator($player, $game);
 
             return response()->json([
                 'status' => 'ok',
@@ -124,7 +126,7 @@ class NegotiatePreContract
 
         // Calculate wage demand (deterministic — no variance)
         $demand = $this->contractService->calculatePreContractWageDemand($player, $this->scoutingService);
-        $mood = $this->getWillingnessMood($player, $game);
+        $mood = $this->dispositionService->willingnessMoodIndicator($player, $game);
         $demandInEuros = (int) ($demand['wage'] / 100);
 
         return response()->json([
@@ -208,7 +210,7 @@ class NegotiatePreContract
                         ]),
                         'wage' => (int) ($offer->wage_counter_offer / 100),
                         'years' => $offer->preferred_years,
-                        'mood' => $this->getWillingnessMood($player, $game),
+                        'mood' => $this->dispositionService->willingnessMoodIndicator($player, $game),
                     ], [
                         'canAccept' => true,
                         'suggestedWage' => $this->calculateMidpointInEuros($offer->offered_wage, $offer->wage_counter_offer),
@@ -291,18 +293,4 @@ class NegotiatePreContract
         return (int) (ceil(($centsA + $centsB) / 2 / 100 / 10000) * 10000);
     }
 
-    /**
-     * Build the mood indicator from the scout willingness score,
-     * so the negotiation modal matches the scout report.
-     */
-    private function getWillingnessMood(GamePlayer $player, Game $game): array
-    {
-        $willingness = $this->scoutingService->calculateWillingness($player, $game);
-
-        return match ($willingness['label']) {
-            'very_interested', 'open' => ['label' => __('transfers.mood_willing_sign'), 'color' => 'green'],
-            'undecided' => ['label' => __('transfers.mood_open_sign'), 'color' => 'amber'],
-            default => ['label' => __('transfers.mood_reluctant_sign'), 'color' => 'red'],
-        };
-    }
 }
