+++ app/Modules/Transfer/Services/ScoutingService.php
@@ -2,50 +2,20 @@
 
 namespace App\Modules\Transfer\Services;
 
-use App\Models\ClubProfile;
 use App\Models\Game;
 use App\Models\GamePlayer;
 use App\Models\ScoutReport;
 use App\Models\ShortlistedPlayer;
 use App\Models\Team;
-use App\Models\TeamReputation;
 use App\Models\TransferOffer;
 use App\Support\Money;
 use App\Support\PositionMapper;
 use Illuminate\Support\Collection;
 use App\Modules\Player\PlayerAge;
-use App\Modules\Player\Services\PlayerTierService;
 use App\Modules\Transfer\Services\ContractService;
 
 class ScoutingService
 {
-    /**
-     * Acceptance probability modifiers based on reputation gap (source - offering).
-     * Gap ≤ 0 means moving up or lateral → no penalty.
-     */
-    private const REPUTATION_GAP_MODIFIERS = [
-        0 => 1.00,
-        1 => 0.75,
-        2 => 0.45,
-        3 => 0.20,
-        4 => 0.08,
-        5 => 0.02,
-    ];
-
-    /**
-     * Minimum team reputation required for a free agent to accept, based on player tier.
-     * Higher-tier free agents demand higher-reputation clubs.
-     */
-    private const MIN_REPUTATION_BY_PLAYER_TIER = [
-        5 => ClubProfile::REPUTATION_CONTINENTAL,  // €50M+ World Class → need continental+
-        4 => ClubProfile::REPUTATION_ESTABLISHED,   // €20M+ Excellent → need established+
-        3 => ClubProfile::REPUTATION_MODEST,         // €5M+ Good → need modest+
-        2 => ClubProfile::REPUTATION_LOCAL,           // €1M+ Average → any team
-        1 => ClubProfile::REPUTATION_LOCAL,           // <€1M Developing → any team
-    ];
-
-    /** Default modifier for gaps of 5+. */
-    private const REPUTATION_GAP_MAX_MODIFIER = 0.02;
 
     /**
      * Wage premium multipliers for players on expiring contracts (pre-contract signings).
@@ -97,6 +67,7 @@ class ScoutingService
 
     public function __construct(
         private readonly ContractService $contractService,
+        private readonly DispositionService $dispositionService,
     ) {}
 
     /**
@@ -331,36 +302,7 @@ class ScoutingService
      */
     public function calculatePlayerImportance(GamePlayer $player, ?Collection $teammates = null): float
     {
-        // Free agents have no team context — return neutral importance
-        if ($player->team_id === null) {
-            return 0.0;
-        }
-
-        if ($teammates === null) {
-            $teammates = GamePlayer::where('game_id', $player->game_id)
-                ->where('team_id', $player->team_id)
-                ->get();
-        }
-
-        if ($teammates->isEmpty()) {
-            return 0.5;
-        }
-
-        // Rank by overall ability (technical + physical average)
-        $sorted = $teammates->sortByDesc(function ($p) {
-            return ($p->current_technical_ability + $p->current_physical_ability) / 2;
-        })->values();
-
-        $rank = $sorted->search(fn ($p) => $p->id === $player->id);
-
-        if ($rank === false) {
-            return 0.5;
-        }
-
-        // Convert rank to 0.0-1.0 scale (0 = worst, 1 = best)
-        $total = $sorted->count();
-
-        return 1.0 - ($rank / max($total - 1, 1));
+        return $this->dispositionService->playerImportance($player, $teammates);
     }
 
     /**
@@ -551,45 +493,7 @@ class ScoutingService
      */
     public function evaluateLoanRequest(GamePlayer $player, ?Game $game = null): array
     {
-        // Reputation gate: player may refuse to join a lower-reputation club
-        if ($game) {
-            $reputationModifier = $this->calculateReputationModifier($game->team, $player);
-            if ($reputationModifier < 1.0 && rand(1, 100) > (int) ($reputationModifier * 100)) {
-                return [
-                    'result' => 'rejected',
-                    'message' => __('transfers.loan_rejected_not_interested', ['player' => $player->name]),
-                ];
-            }
-        }
-
-        $importance = $this->calculatePlayerImportance($player);
-
-        if ($importance > 0.7) {
-            return [
-                'result' => 'rejected',
-                'message' => __('transfers.loan_rejected_key_player', ['team' => $player->team?->name, 'player' => $player->name]),
-            ];
-        }
-
-        if ($importance > 0.4) {
-            // 50% chance
-            if (rand(0, 1) === 1) {
-                return [
-                    'result' => 'accepted',
-                    'message' => __('transfers.loan_accepted', ['team' => $player->team?->name, 'player' => $player->name]),
-                ];
-            }
-
-            return [
-                'result' => 'rejected',
-                'message' => __('transfers.loan_rejected_keep', ['team' => $player->team?->name, 'player' => $player->name]),
-            ];
-        }
-
-        return [
-            'result' => 'accepted',
-            'message' => __('transfers.loan_accepted', ['team' => $player->team?->name, 'player' => $player->name]),
-        ];
+        return $this->dispositionService->evaluateLoanRequest($player, $game);
     }
 
     // =========================================
@@ -600,52 +504,11 @@ class ScoutingService
      * Deterministic loan request evaluation for sync negotiation.
      * Returns result, asking loan fee, mood, and rejection reason.
      *
-     * @return array{result: string, loan_fee: int, disposition: float, rejection_reason: ?string}
+     * @return array{result: string, disposition: float, rejection_reason: ?string}
      */
     public function evaluateLoanRequestSync(GamePlayer $player, Game $game): array
     {
-        // Gate 1: Reputation — club won't negotiate with low-rep teams
-        $reputationModifier = $this->calculateReputationModifier($game->team, $player);
-        if ($reputationModifier < 0.50) {
-            return [
-                'result' => 'rejected',
-                'disposition' => 0.10,
-                'rejection_reason' => 'reputation',
-            ];
-        }
-
-        $importance = $this->calculatePlayerImportance($player);
-
-        // Gate 1: Key player — club refuses to loan
-        if ($importance > 0.70) {
-            return [
-                'result' => 'rejected',
-                'disposition' => 0.15,
-                'rejection_reason' => 'key_player',
-            ];
-        }
-
-        // Calculate disposition for mood indicator
-        $disposition = 0.50;
-        $disposition += (1.0 - $importance) * 0.30;
-        $disposition += ($reputationModifier - 0.50) * 0.20;
-        $disposition = max(0.10, min(0.95, $disposition));
-
-        // Gate 2: Player willingness — player may not want to join
-        $willingness = $this->calculateWillingness($player, $game, $importance);
-        if (in_array($willingness['label'], ['not_interested', 'reluctant'])) {
-            return [
-                'result' => 'rejected',
-                'disposition' => $disposition,
-                'rejection_reason' => 'player_refused',
-            ];
-        }
-
-        return [
-            'result' => 'accepted',
-            'disposition' => $disposition,
-            'rejection_reason' => null,
-        ];
+        return $this->dispositionService->evaluateLoanRequestSync($player, $game);
     }
 
     // =========================================
@@ -709,29 +572,7 @@ class ScoutingService
      */
     public function calculateReputationModifier(Team $biddingTeam, GamePlayer $player): float
     {
-        // Free agents have no current team context — no penalty
-        if ($player->team_id === null) {
-            return 1.0;
-        }
-
-        $gameId = $player->game_id;
-        $sourceReputation = $gameId && $player->team_id
-            ? TeamReputation::resolveLevel($gameId, $player->team_id)
-            : ($player->team?->clubProfile?->reputation_level ?? ClubProfile::REPUTATION_LOCAL);
-        $offeringReputation = $gameId
-            ? TeamReputation::resolveLevel($gameId, $biddingTeam->id)
-            : ($biddingTeam->clubProfile?->reputation_level ?? ClubProfile::REPUTATION_LOCAL);
-
-        $sourceIndex = ClubProfile::getReputationTierIndex($sourceReputation);
-        $offeringIndex = ClubProfile::getReputationTierIndex($offeringReputation);
-
-        $gap = $sourceIndex - $offeringIndex;
-
-        if ($gap <= 0) {
-            return 1.0; // Moving up or lateral
-        }
-
-        return self::REPUTATION_GAP_MODIFIERS[$gap] ?? self::REPUTATION_GAP_MAX_MODIFIER;
+        return $this->dispositionService->reputationModifier($biddingTeam, $player);
     }
 
     // =========================================
@@ -744,7 +585,7 @@ class ScoutingService
      */
     public function canSignFreeAgent(GamePlayer $player, string $gameId, string $teamId): bool
     {
-        return $this->getFreeAgentWillingnessLevel($player, $gameId, $teamId) === 'willing';
+        return $this->dispositionService->canSignFreeAgent($player, $gameId, $teamId);
     }
 
     /**
@@ -754,25 +595,7 @@ class ScoutingService
      */
     public function getFreeAgentWillingnessLevel(GamePlayer $player, string $gameId, string $teamId): string
     {
-        $playerTier = $player->tier ?? PlayerTierService::tierFromMarketValue($player->market_value_cents);
-        $minReputation = self::MIN_REPUTATION_BY_PLAYER_TIER[$playerTier] ?? ClubProfile::REPUTATION_LOCAL;
-
-        $teamReputation = TeamReputation::resolveLevel($gameId, $teamId);
-
-        $teamIndex = ClubProfile::getReputationTierIndex($teamReputation);
-        $minIndex = ClubProfile::getReputationTierIndex($minReputation);
-
-        $gap = $minIndex - $teamIndex;
-
-        if ($gap <= 0) {
-            return 'willing';
-        }
-
-        if ($gap === 1) {
-            return 'reluctant';
-        }
-
-        return 'unwilling';
+        return $this->dispositionService->freeAgentWillingnessLevel($player, $gameId, $teamId);
     }
 
     // =========================================
@@ -982,64 +805,7 @@ class ScoutingService
      */
     public function calculateWillingness(GamePlayer $player, Game $game, ?float $importance = null): array
     {
-        $importance ??= $this->calculatePlayerImportance($player);
-
-        // Base willingness: low importance players are more willing
-        $score = (int) ((1.0 - $importance) * 50);
-
-        // Contract length factor: fewer years left = more willing
-        if ($player->contract_until) {
-            $yearsLeft = max(0, $game->current_date->diffInYears($player->contract_until));
-            if ($yearsLeft <= 1) {
-                $score += 30;
-            } elseif ($yearsLeft <= 2) {
-                $score += 15;
-            }
-        } else {
-            $score += 25; // No contract = very willing
-        }
-
-        // Age factor: older players at lower-rep clubs more open
-        $age = $player->age($game->current_date);
-        if ($age >= PlayerAge::PRIME_END) {
-            $score += 10;
-        } elseif ($age < PlayerAge::YOUNG_END) {
-            $score += 5; // Young players seeking opportunities
-        }
-
-        // Reputation gap: penalize moving down, reward moving up
-        $reputationModifier = $this->calculateReputationModifier($game->team, $player);
-        if ($reputationModifier < 1.0) {
-            // Moving down: scale the score down proportionally to the reputation gap
-            $score = (int) ($score * $reputationModifier);
-        } elseif ($player->team_id) {
-            // Moving up: bonus based on how many tiers above the buying club is
-            $sourceReputation = TeamReputation::resolveLevel($player->game_id, $player->team_id);
-            $offeringReputation = TeamReputation::resolveLevel($player->game_id, $game->team_id);
-            $sourceIndex = ClubProfile::getReputationTierIndex($sourceReputation);
-            $offeringIndex = ClubProfile::getReputationTierIndex($offeringReputation);
-            $upwardGap = $offeringIndex - $sourceIndex;
-
-            if ($upwardGap >= 3) {
-                $score += 30; // Dream move (e.g. local → elite)
-            } elseif ($upwardGap === 2) {
-                $score += 20; // Big step up
-            } elseif ($upwardGap === 1) {
-                $score += 10; // Step up
-            }
-        }
-
-        $score = min(100, max(0, $score + rand(-5, 5)));
-
-        $label = match (true) {
-            $score >= 80 => 'very_interested',
-            $score >= 60 => 'open',
-            $score >= 40 => 'undecided',
-            $score >= 20 => 'reluctant',
-            default => 'not_interested',
-        };
-
-        return ['score' => $score, 'label' => $label];
+        return $this->dispositionService->playerTransferWillingness($player, $game, $importance);
     }
 
     /**
